# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EP1 Exam",
    "description": "documentation for Noroff EP - e-commerce"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets list of all products",
        "produces": [
          "json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Creates new product",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/postproduct"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Alters existing product",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/alterproduct"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Soft deletes product of provided id(sets Processed : 1)",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{productid}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets product of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "productid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/{userid}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets user of provided id",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "userid",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Gets list of users",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Alters user",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/alteruser"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Delete user",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/delete"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/membership": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Alter users membership manually",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/usermembership"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/role": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Alter users role manually",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/userrole"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "$ref": "#/security/Admin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "delete": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "id"
      ]
    },
...
    "postproduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Season tickets"
        },
        "description": {
          "type": "string",
          "example": "Hockey tickets for the NOSHO season"
        },
        "price": {
          "type": "number",
          "example": 499
        },
        "quantity": {
          "type": "number",
          "example": 20
        },
        "imgurl": {
          "type": "string",
          "example": "https://i0.wp.com/puckjunk.com/wp-content/uploads/2024/05/Shoresy-Poster.jpg?ssl=1"
        }
      },
      "required": [
        "name",
        "description",
        "price",
        "quantity",
        "imgurl"
      ]
    },
    "alterproduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Season tickets"
        },
        "description": {
          "type": "string",
          "example": "Hockey tickets for the NOSHO season"
        },
        "price": {
          "type": "number",
          "example": 499
        },
        "quantity": {
          "type": "number",
          "example": 20
        },
        "imgurl": {
          "type": "string",
          "example": "https://i0.wp.com/puckjunk.com/wp-content/uploads/2024/05/Shoresy-Poster.jpg?ssl=1"
        }
      },
      "required": [
        "id",
        "name",
        "description",
        "price",
        "quantity",
        "imgurl"
      ]
    },
...
    "alteruser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 2
        },
        "firstName": {
          "type": "string",
          "example": "Shoresy"
        },
        "lastName": {
          "type": "string",
          "example": "Waffles"
        },
        "phone": {
          "type": "number",
          "example": 69
        },
        "userName": {
          "type": "string",
          "example": "neverLoose"
        },
        "address": {
          "type": "string",
          "example": "The ice rank 69"
        }
      },
      "required": [
        "id",
        "firstName",
        "lastName",
        "phone",
        "userName",
        "address"
      ]
    },
    "usermembership": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "number",
          "example": 2
        },
        "membershipid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "userid",
        "membershipid"
      ]
    },
    "userrole": {
      "type": "object",
      "properties": {
        "userid": {
          "type": "number",
          "example": 2
        },
        "roleid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "userid",
        "roleid"
      ]
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "token",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": {
    "Admin": [
      "adminJWTtoken"
    ]
  }
}